split image upload/creation. clean up
